feat: implement ability to skip package creation if there are no changes
Added diffcheck flag, to package creation tasks, that causes the package to only
build if there are changes to the package. A change is considered to have
occurred if package files have been modified between HEAD and the last build
tag.

Added gittag flag, to package creation tasks, that if enabled pushes a tag to
the repository upon successful build.