Introduce an Install Source Package Task
Install Source Package task simplifies the use of source packages,
which currently rely on checkout and deploy. It also brings in
features when used in Azure Pipelines such as the ability to skip,
if the package is already deployed in the org.

This implies Create Source Package has some changes to metadata as
well as to the artifacts being produced, but it will still be
backward compatbile with Checkout and Deploy Tasks.