Remove the need for type of artifact and alias from Install/Checkout Tasks (#151)
This enhancement removes the need for asking the user for artifact type in Azure Pipelines. Classic Release pipelines and new YAML pipelines unpacks attached artifacts into a default artifact directory. This can be utilized to understand the artifact locations automatically by globing this directory.

Moreover, the enhancement also has an option for alternate artifact download directories, which could be anything the user is pointing to.