Refactor PackageManifest class
Change method of instantiation of PackageManifest to factory method. This allows
the package.xml to be parsed when PackageManifest is instantiated, as
constructors cannot be asynchronous. By enforcing the xml to be parsed at object
creation, the instance methods are guaranteed to have access to the manifest.

Fix bug where 'types' is undefined and causes an error trying to reference a
property on type undefined.