fix(source-tracking): clean up mpd to prevent source tracking lib to gather status of all files
Possibly source tracking lib use statusMatrix to gain information about all files when the
repository is initiated. This is gathered from the projects path at the point when the library is
intiatlized. As sfpowerscripts fixes source tracking one after the another, this is not necessary
and is resulting in some errors (statusMatrix unable to find files.. though no functional impact)
and also unnecessary computation. This fix is to clean up the sfdx-project.json to only carry the
package that is currently being fixed up for source tracking