feat(fast-feedback): fast feedback to skip packages from validation when descriptor changes
validate will validate the full package when the descriptor of the package changes. Most often the
change is in a version number, and this triggers full deployment and testing, in fast feedback mode,
this could be skipped. This has a negative issue when additonal orchestrator parameters are added
like assignPermissionSets etc, this could be validated in a thorough mode