feat(prepare): add ability to daisy chain prepare command
prepare can take extremely long hours on some repos often running out of time in hosted ci agents
which are set at 3 or 6 hours depending on the platform. This allows to daisychain prepare command
where a pool prepared in a job can be used by another job. This is introduced by adding an
additional flag called snapshotPool in the pool config. (alpha feature)