feat(sfp-package): add support for unpackaged metadata in unlocked packages
Add support for unpackaged metadata. This was removed as sfpowerscripts do parallel development and
working directory didnt copy the unpackaged directory. Support is added to copy unpackaged directory
and copy the folder to appropriate working directory. sfdx-project.json is then aligned for
sucessful preparation of the package. The unpackaged is a package directory and it will be upto the
user to determine the behaviour during validate,prepare or deploy. This along with changes in #842
which will skip the directories

fix #958