feat(changelog): add support for multiple workitems filter in changelog:generate and release
This feature adds support for multiple workitems filter, where in the release defn one could add
different filters. This means the YAML schema need to be updated    d to input an array instead of
string

BREAKING CHANGE: the schema of release definition is breaking, as it now supports an array of
workitem filters as opposed to a single value. This means users should update their release
definition