fix(release-defn): ensure release name is a mandatory flag
Release Name in yaml makes it hard to modify, changelogbranch ref is incorrect semnatically and very
restrictive from a usage perspective. Hence both of this is removed. Instead releasename is enforced
as a mandatory flag

BREAKING CHANGE: ReleaseName and ChangelogBranchRef in YAML is removed, only supports for release
name passed in