fix(publish): fix publish adding an additional '@'
Changes in @salesforce/command has led to ignoring the removal of '@' symbol added by a user during
the parsing of the flag. This has led subsequent code to add '@' again, as '@' is mandatory for
scopde package and users may ignore adding '@' to the scope. This was clearly the established
behaviour with publish command. This fix reintroduces previous behaviours

issue not created in github, but refered in slack as
https://dxatscale.slack.com/archives/CN9N54GF5/p1670321226990929