feat(validate): display info about the org that is used for validation (#1239)
* feat(validate): display info about the org that is used for validation

In some cases devs would like to understand what went wrong in a validation run. sfpowerscripts
currently do not provide an easy mechanism which displays those details. So in those instances
additional scripts need to be done that will display the org info by traversing the org list. This
enhancement will display the org info when an additional flag is passed to validate.