feat: add a 'diff' package type (#1320)
This functionality adds a mechanism to migrate to DX@Scale project more easily. A new package type 'diff'
provides a mechanism where a package contains only metadata that is baselined against the commit id in
production against the incoming version. This allows a current project which is using sandboxes to migrate
to modules along with package types in the repository. Any packages below this package need to be
source packages and should assume the target org will have the necessary metadata.