fix(build): do not fail if there is nothing to be built
When used across matrix job in CI platforms, there are instances when build gets triggered but on
deep introspection, there will be nothing to be built. However the CI platforms have different
intricacies which make dependent jobs really difficult. To simplify this, we will not let build
failwhe

BREAKING CHANGE: Build will not fail if there is nothing to be built